Build fixes
