+++ apiserver/subnets/subnets.go
@@ -4,8 +4,6 @@
 package subnets
 
 import (
-	"strings"
-
 	"github.com/juju/errors"
 	"github.com/juju/loggo"
 	"github.com/juju/names"
@@ -13,8 +11,6 @@ import (
 	"github.com/juju/juju/apiserver/common"
 	"github.com/juju/juju/apiserver/common/networkingcommon"
 	"github.com/juju/juju/apiserver/params"
-	"github.com/juju/juju/environs"
-	providercommon "github.com/juju/juju/provider/common"
 	"github.com/juju/juju/state"
 )
 
@@ -24,7 +20,25 @@ func init() {
 	common.RegisterStandardFacade("Subnets", 1, NewAPI)
 }
 
-// subnetsAPI implements the networkingcommon.SubnetsAPI interface.
+// SubnetsAPI defines the methods the Subnets API facade implements.
+type SubnetsAPI interface {
+	// AllZones returns all availability zones known to Juju. If a
+	// zone is unusable, unavailable, or deprecated the Available
+	// field will be false.
+	AllZones() (params.ZoneResults, error)
+
+	// AllSpaces returns the tags of all network spaces known to Juju.
+	AllSpaces() (params.SpaceResults, error)
+
+	// AddSubnets adds existing subnets to Juju.
+	AddSubnets(args params.AddSubnetsParams) (params.ErrorResults, error)
+
+	// ListSubnets returns the matching subnets after applying
+	// optional filters.
+	ListSubnets(args params.SubnetsFilters) (params.ListSubnetsResults, error)
+}
+
+// subnetsAPI implements the SubnetsAPI interface.
 type subnetsAPI struct {
 	backing    networkingcommon.NetworkBacking
 	resources  *common.Resources
@@ -33,13 +47,13 @@ type subnetsAPI struct {
 
 // NewAPI creates a new Subnets API server-side facade with a
 // state.State backing.
-func NewAPI(st *state.State, res *common.Resources, auth common.Authorizer) (networkingcommon.SubnetsAPI, error) {
+func NewAPI(st *state.State, res *common.Resources, auth common.Authorizer) (SubnetsAPI, error) {
 	return newAPIWithBacking(networkingcommon.NewStateShim(st), res, auth)
 }
 
 // newAPIWithBacking creates a new server-side Subnets API facade with
 // a common.NetworkBacking
-func newAPIWithBacking(backing networkingcommon.NetworkBacking, resources *common.Resources, authorizer common.Authorizer) (networkingcommon.SubnetsAPI, error) {
+func newAPIWithBacking(backing networkingcommon.NetworkBacking, resources *common.Resources, authorizer common.Authorizer) (SubnetsAPI, error) {
 	// Only clients can access the Subnets facade.
 	if !authorizer.AuthClient() {
 		return nil, common.ErrPerm
@@ -53,42 +67,7 @@ func newAPIWithBacking(backing networkingcommon.NetworkBacking, resources *commo
 
 // AllZones is defined on the API interface.
 func (api *subnetsAPI) AllZones() (params.ZoneResults, error) {
-	var results params.ZoneResults
-
-	zonesAsString := func(zones []providercommon.AvailabilityZone) string {
-		results := make([]string, len(zones))
-		for i, zone := range zones {
-			results[i] = zone.Name()
-		}
-		return `"` + strings.Join(results, `", "`) + `"`
-	}
-
-	// Try fetching cached zones first.
-	zones, err := api.backing.AvailabilityZones()
-	if err != nil {
-		return results, errors.Trace(err)
-	}
-
-	if len(zones) == 0 {
-		// This is likely the first time we're called.
-		// Fetch all zones from the provider and update.
-		zones, err = api.updateZones()
-		if err != nil {
-			return results, errors.Annotate(err, "cannot update known zones")
-		}
-		logger.Debugf(
-			"updated the list of known zones from the environment: %s", zonesAsString(zones),
-		)
-	} else {
-		logger.Debugf("using cached list of known zones: %s", zonesAsString(zones))
-	}
-
-	results.Results = make([]params.ZoneResult, len(zones))
-	for i, zone := range zones {
-		results.Results[i].Name = zone.Name()
-		results.Results[i].Available = zone.Available()
-	}
-	return results, nil
+	return networkingcommon.AllZones(api.backing)
 }
 
 // AllSpaces is defined on the API interface.
@@ -110,57 +89,13 @@ func (api *subnetsAPI) AllSpaces() (params.SpaceResults, error) {
 	return results, nil
 }
 
-// zonedEnviron returns a providercommon.ZonedEnviron instance from
-// the current environment config. If the environment does not support
-// zones, an error satisfying errors.IsNotSupported() will be
-// returned.
-func (api *subnetsAPI) zonedEnviron() (providercommon.ZonedEnviron, error) {
-	envConfig, err := api.backing.EnvironConfig()
-	if err != nil {
-		return nil, errors.Annotate(err, "getting environment config")
-	}
-
-	env, err := environs.New(envConfig)
-	if err != nil {
-		return nil, errors.Annotate(err, "opening environment")
-	}
-	if zonedEnv, ok := env.(providercommon.ZonedEnviron); ok {
-		return zonedEnv, nil
-	}
-	return nil, errors.NotSupportedf("availability zones")
-}
-
-// updateZones attempts to retrieve all availability zones from the
-// environment provider (if supported) and then updates the persisted
-// list of zones in state, returning them as well on success.
-func (api *subnetsAPI) updateZones() ([]providercommon.AvailabilityZone, error) {
-	zoned, err := api.zonedEnviron()
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	zones, err := zoned.AvailabilityZones()
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	if err := api.backing.SetAvailabilityZones(zones); err != nil {
-		return nil, errors.Trace(err)
-	}
-	return zones, nil
-}
-
 // AddSubnets is defined on the API interface.
 func (api *subnetsAPI) AddSubnets(args params.AddSubnetsParams) (params.ErrorResults, error) {
-	return networkingcommon.AddSubnets(api, args)
+	return networkingcommon.AddSubnets(api.backing, args)
 }
 
 // ListSubnets lists all the available subnets or only those matching
 // all given optional filters.
 func (api *subnetsAPI) ListSubnets(args params.SubnetsFilters) (results params.ListSubnetsResults, err error) {
-	return networkingcommon.ListSubnets(api, args)
-}
-
-// Backing is defined on the API interface.
-func (api *subnetsAPI) Backing() networkingcommon.NetworkBacking {
-	return api.backing
+	return networkingcommon.ListSubnets(api.backing, args)
 }
